Update model validations to pass

+++ db/migrate/20201017195306_create_items.rb
@@ -3,7 +3,7 @@ class CreateItems < ActiveRecord::Migration[5.2]
     create_table :items do |t|
       t.string :name
       t.string :description
-      t.integer :unit_price
+      t.float :unit_price
       t.string :merchant_id, foreign_key: true
       t.datetime :created_at
       t.datetime :updated_at
