+++ db/migrate/20201018012925_create_invoice_items.rb
@@ -4,7 +4,7 @@ class CreateInvoiceItems < ActiveRecord::Migration[5.2]
       t.string :item_id, foreign_key: true
       t.string :invoice_id, foreign_key: true
       t.integer :quantity
-      t.integer :unit_price
+      t.float :unit_price
       t.datetime :created_at
       t.datetime :updated_at
